chapter5/cryptocurrency_application/requirements.txt
sanic==0.7.0
pycryptodome==3.4.11
ecdsa==0.13
